logstash-mutual-tls/requirements.txt
colorama==0.4.4
Jinja2==2.11.3
pyyaml==5.4.1
requests==2.25.1
